test: split validation tests to match GROW/POLISH module split
Move 76 passage-layer check tests to new test_polish_passage_validation.py.
Keep 34 beat-DAG check tests in test_grow_validation.py. New file imports
from polish_validation directly rather than via backward-compat re-exports.
